+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## v1.1 
+
+### Added
+
+- Ban and kick (working on mute)
+
 ## v1 - The "Social" Update
 
 - We are out of beta, we're releasing on time (If you didn't get that's a Portal 2 reference :3)